+++ src/Sentry.AspNet/SentryAspNetOptionsExtensions.cs
@@ -1,3 +1,4 @@
+using System.Web;
 using Sentry.AspNet.Internal;
 using Sentry.Extensibility;
 using Sentry.Infrastructure;
@@ -23,7 +24,7 @@ namespace Sentry.AspNet
             var eventProcessor = new SystemWebRequestEventProcessor(payloadExtractor, options);
 
             options.DiagnosticLogger ??= new DebugDiagnosticLogger(options.DiagnosticLevel);
-            options.Release ??= SystemWebVersionLocator.GetCurrent();
+            options.Release ??= SystemWebVersionLocator.Resolve(options, HttpContext.Current);
             options.AddEventProcessor(eventProcessor);
         }
     }