+++ src/Sentry/ISentryClient.cs
@@ -38,6 +38,16 @@ namespace Sentry
         void CaptureTransaction(Transaction transaction);
 
         /// <summary>
+        /// Captures a session update.
+        /// </summary>
+        /// <remarks>
+        /// Note: this method is NOT meant to be called from user code!
+        /// It will be called automatically by the SDK.
+        /// </remarks>
+        /// <param name="sessionUpdate">The update to send to Sentry.</param>
+        void CaptureSession(SessionUpdate sessionUpdate);
+
+        /// <summary>
         /// Flushes events queued up.
         /// </summary>
         /// <param name="timeout">How long to wait for flush to finish.</param>