+++ src/Sentry/ISessionManager.cs
@@ -1,0 +1,11 @@
+﻿namespace Sentry
+{
+    internal interface ISessionManager
+    {
+        SessionUpdate? StartSession();
+
+        SessionUpdate? ReportError();
+
+        SessionUpdate? EndSession(SessionEndStatus status);
+    }
+}